CLOSED: Link posts to users

+++ blog/app/Http/Controllers/PostController.php
@@ -13,9 +13,12 @@ use Stevebauman\Purify\Facades\Purify;
 use Intervention\Image\ImageManager;
 use Intervention\Image\Drivers\GD\Driver;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
+
     /**
      * Display a listing of the resource.
      */
@@ -56,6 +59,7 @@ class PostController extends Controller
         $post->title = $request->title;
         $post->slug = $request->slug;
         $post->category_id = $request->category_id;
+        $post->user_id = Auth::user()->id;
         $post->body = Purify::clean($request->body); // Purify dirty html
 
         // Save Image
@@ -102,6 +106,14 @@ class PostController extends Controller
         // Find the post in the db and save it as a variable
         $post = Post::find($id);
 
+        // Authorize the user
+        $response = Gate::inspect('update', $post);
+
+        if(!$response->allowed()){
+            Session::flash('error', $response->message());
+            return back();
+        }
+
         $categories = Category::all();
         $tags = Tag::all();
         // return the view
@@ -114,7 +126,16 @@ class PostController extends Controller
     public function update(Request $request, string $id)
     {
         // Validate the data
-        $post = Post::find($id);
+        $post = Post::findOrFail($id);
+
+        // Authorize the user
+        $response = Gate::inspect('update', $post);
+
+        if(!$response->allowed()){
+            Session::flash('error', $response->message());
+            return back();
+        }
+
         $request->validate([
             'title' => 'required|max:255',
             'slug' => "required|alpha_dash|min:5|max:255|unique:posts,slug,$id",
@@ -127,6 +148,7 @@ class PostController extends Controller
         $post->title = $request->title;
         $post->slug = $request->slug;
         $post->category_id = $request->category_id;
+        $post->user_id = Auth::user()->id;
         $post->body = Purify::clean($request->body);// Purify dirty html
 
         // Save Image
@@ -171,7 +193,15 @@ class PostController extends Controller
      */
     public function destroy(string $id)
     {
-        $post = Post::find($id);
+        $post = Post::findOrFail($id);
+
+        // Authorize the user
+        $response = Gate::inspect('delete', $post);
+
+        if(!$response->allowed()){
+            Session::flash('error', $response->message());
+            return back();
+        }
 
         $post->delete();
